requirements.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.115.11
greenlet==3.1.1
h11==0.14.0
idna==3.10
packaging==24.2
pydantic==2.10.6
pydantic_core==2.27.2
pyhumps==3.8.0
PyYAML==6.0.2
setuptools-scm==8.2.0
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
toml==0.10.2
tomli==2.2.1
typing_extensions==4.12.2
uvicorn==0.34.0
